# Conflicts:
#	target/classes/META-INF/maven/no.ntnu.unnamedsoftware/russesamfunnet-rest/pom.properties

+++ src/main/java/no/ntnu/unnamedsoftware/service/FeedService.java
@@ -15,7 +15,7 @@ public class FeedService {
 	
 	public String getSchoolFeed(int theRussId)
 	{
-		int theSchoolId = feedDAO.getSchoolId(theRussId);
+		Long theSchoolId = feedDAO.getSchoolId(theRussId);
 		return feedDAO.getSchoolFeed(theSchoolId);
 	}
